fix: error checking — reproducibility, batch size, test coverage
- Sort candidates in build_haystacked_a and build_dilution_control
  for deterministic hay selection (was dependent on dict iteration order)
- Reduce OpenAI embedding batch size from 2048 to 100 to avoid
  token limit errors with long email bodies
- Wrap numpy float in displacement() for JSON serialization safety
- Clarify direction convention in compare_conditions docstring
- Add test for build_haystacked_c (embedding-optimized condition)
- Add tests for run_full_experiment and save_results

+++ src/metrics.py
@@ -35,7 +35,7 @@ def displacement(
         attack_rank = _rank_of(attack_ranked, doc_id)
         displacements.append(attack_rank - base_rank)
 
-    return np.mean(displacements) if displacements else 0.0
+    return float(np.mean(displacements)) if displacements else 0.0
 
 
 def bootstrap_ci(
